+++ docs/typescript/api-reference/overview.mdx
@@ -110,7 +110,7 @@ Create a server with `new MCPServer({ ... })`. See the [MCPServer](/typescript/a
     icon="list-checks"
     href="/typescript/api-reference/server/elicitation-schemas"
   >
-    Enum and form schema builders for elicitation.
+    Migrate v1 enum elicitation helpers to Zod.
   </Card>
   <Card
     title="Authentication"
