+++ libraries/typescript/packages/server/tsup.config.ts
@@ -54,9 +54,9 @@ export default defineConfig([
       options.minifyIdentifiers = false;
     },
   },
-  // Deprecated temporary v1 compatibility entry. Build it independently so
-  // shared-chunk extraction cannot alter the native root entry's graph or
-  // edge/startup budgets. Remove this whole target in mcp-use v3.
+  // Deprecated v1 bridge. Uses splitting so compat-v1.js stays a thin entry, but
+  // remains a separate target so the native index graph is unchanged. Minified
+  // post-build. Removed in mcp-use v3.
   {
     entry: {
       "compat-v1": "src/compat-v1.ts",
@@ -64,7 +64,7 @@ export default defineConfig([
     format: ["esm"],
     target: "node22",
     dts: false,
-    splitting: false,
+    splitting: true,
     sourcemap: false,
     clean: false,
     external: ["@mcp-use/client", "@mcp-use/cli"],
